requirements.txt
attrs==23.1.0
certifi==2023.11.17
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
pip==23.3.2
PySocks==1.7.1
python-dotenv==1.0.0
selenium==4.16.0
setuptools==68.2.2
slack-sdk==3.26.1
sniffio==1.3.0
sortedcontainers==2.4.0
trio-websocket==0.11.1
trio==0.23.2
urllib3==1.26.6
wheel==0.41.2
wsproto==1.2.0
